scripts/requirements.txt
azure-ai-formrecognizer
azure-search-documents
azure-identity
openai
python-dotenv
